requirements.txt
streamlit
PyQt5
PyQtWebEngine
requests
Pillow
